backend/pyproject.toml
[tool.poetry]
name = "iWitness Backend"
version = "0.1.0"
description = "Python backend project with FastAPI"
authors = ["Eternal-Engine <ninovationlab@gmail.com>"]
readme = "BACKEND.md"

[tool.poetry.dependencies]
aiosql = "^3.4.1"
asyncpg = "^0.25.0"
fastapi = {version = "^0.75.1"}
httpx = "^0.22.0"
ipython = "^8.2.0"
loguru = "^0.6.0"
python = "^3.10"
pydantic = {extras = ["email", "dotenv"], version = "^1.9.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
psycopg2-binary = "^2.9.3"
python-slugify = "^6.1.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
SQLAlchemy = "^1.4.32"
trio = "^0.20.0"
orjson = "^3.6.7"
requests = "^2.27.1"
asgi-lifespan = "^1.0.1"
argon2-cffi = "^21.3.0"
bcrypt = "^3.2.0"
jwt = "^1.3.1"
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
PyPika = "^0.48.9"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]
alembic = "^1.7.7"
bandit = "^1.7.4"
black = "^22.1.0"
colorama = "^0.4.4"
flake8 = "^3.9.2"
isort = "^5.10.1"
mypy = "^0.941"
pre-commit = "^2.17.0"
pycodestyle = "^2.7.0"
pylint = "^2.12.2"
pyproject-flake8 = "0.0.1a2"
pytest = "^7.1.1"
pytest-asyncio = "^0.18.2"
pytest-cov = "^3.0.0"
pytest-xdist = "^2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
include = '\.pyi?$'
exclude = '''
  /(
      \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist

    # The following are specific to Black, you probably don't want those.
    | blib2to3
    | tests/data
  )/
'''

[tool.isort]
combine_as_imports = true
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 119
multi_line_output = 3
profile = "black"
src_paths = ["app/", "tests/"]
skip = [
  ".gitignore",
  ".dockerignore",
  ".env",
  "coverage/*",
  ".html",
  ".xml",
  ".github/",
  ".rst",
  ".md",
]
use_parentheses = true

[tool.pytest.ini_options]
testpaths = "tests"
filterwarnings = "error"
addopts = '''
  --strict-markers
  --tb=short
  --cov=app
  --cov=tests
  --cov-branch
  --cov-report=term-missing
  --cov-report=html:coverage/cov_html
  --cov-report=xml:coverage/cov.xml
  --no-cov-on-fail
  --cov-fail-under=90
  --numprocesses=auto
  --asyncio-mode=auto
'''

[tool.coverage.run]
omit = ["app/db/database.py"]

[tool.pylint.master]
max-line-length = 119
ignore-paths = ["./tests/"]

[tool.pylint.messages_control]
disable = [
  "C0103", "C0114", "C0116", "C0115", "C0330", "C0326", "C0411", "C0415",
  "E1101", "E1111", "E0213", "E0401", "E0611", "E1121",
  "R0201", "R0801", "R0903",
  "W0106", "W0143", "W0613", "W0621", "W0622", "W1203",
  "E0211", "E1123" # > TOBE DELETED FOR DEPLOYMENT
]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict_optional = false
warn_no_return = true
warn_return_any = true
pretty = true
color_output = true
error_summary = true

[tool.flake8]
max-line-length = 119
exclude = ["tests/*"]
extend-ignore = "E1,E23,E203,E501,W503"
max-complexity = 10
